+++ effectful/CHANGELOG.md
@@ -7,6 +7,8 @@
   `Provider` effect.
 * Add `runErrorWith` and `runErrorNoCallStackWith` to `Effectful.Error.Dynamic`
   and `Effectful.Error.Static`.
+* Add support for having multiple effects of the same type in scope via the
+  `Labeled` effect.
 
 # effectful-2.2.2.0 (2023-01-11)
 * Add `withSeqEffToIO` and `withConcEffToIO` to `Effectful`.